kernel/platform: Implement mmio_read and mmio_write for native platform
Introduce `mmio_write` and `mmio_read` for native platform. This is useful
when developing or testing drivers that rely on MMIO, like virtio.

It was not possible to implement these two function using generics
because SvsmPlatform struct uses the `dyn` trait.

It is also not possible to iterate and read byte by byte as this
would violate the virtio spec, where is stated that all multi-byte
long registers must be read in one operation, therefore making it
impossible to implement a virtio driver.

Signed-off-by: Luigi Leonardi <leonardi@redhat.com>