kernel/vsock: send credit update
virtio-vsock spec forbids to send data if the receiver's queue is full.
If the guest is expecting some more data from the host and the host is waiting
for some space in the buffer, this leads to a deadlock.

Send a credit update to inform the peer about the space available in the queue.

Signed-off-by: Luigi Leonardi <leonardi@redhat.com>