github/workflows: enable vhost support in QEMU build and install netcat
In preparation to vsock support, QEMU needs to be configured with vhost
support enabled. Without the flag `--enable-vhost-kernel`, QEMU would fail
to launch with error:
`qemu-system-x86_64: -device vhost-vsock-device,guest-cid=3: 'vhost-vsock-device' is not a valid device model name`

This flag is disabled by default because of `--without-default-features`

Because no-cc is run as non-root, this causes QEMU to fail with the
error: `vhost-vsock: failed to open vhost device: Permission denied"`

To fix this, setup permissions to /dev/vhost-vsock for non-root.

Add the flag to enable support for vsock in QEMU and install netcat

Co-developed-by: Luigi Leonardi <leonardi@redhat.com>
Signed-off-by: Luigi Leonardi <leonardi@redhat.com>
Signed-off-by: Nicola Ramacciotti <niko.ramak@gmail.com>