kernel/block: Track virtio-MMIO slots status
Currently the virtio-blk driver probes every virtio-mmio address, regardless of
its status. This is problematic, when more than one driver is loaded, (eg virtio-blk
and virtio-vsock) because the probe function will reset the device, if the device
type is not the one the driver is looking for.

Introduce a static array that keeps track of the status of each slot
and update the probing mechanism in the driver, so that just free
slots are considered.

Signed-off-by: Luigi Leonardi <leonardi@redhat.com>